backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
opencv-python
pillow
sentence-transformers
google-api-python-client
numpy
pandas
ultralytics
seaborn
matplotlib
pyyaml
requests
tqdm
scipy
python-multipart
google-generativeai
python-dotenv
youtube-search-python
moviepy
edge-tts
